requirements_production.txt
# Production Requirements for Enhanced Whitelist + Resource Limits v2.0
# ==================================================================
# 
# Optimized for VPS deployment with enterprise-grade security,
# performance, and monitoring capabilities.
#
# Usage:
#   pip install -r requirements_production.txt

# ================================
# 🌐 CORE FLASK & WEB FRAMEWORK
# ================================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# ================================  
# 🔐 AUTHENTICATION & SECURITY
# ================================
Flask-Login==0.6.3
Flask-Dance==7.0.0
requests-oauthlib==1.3.1
authlib==1.2.1

# Security Headers & CSRF Protection
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
WTForms==3.1.0

# Password & Encryption
cryptography==41.0.7
bcrypt==4.1.1

# ================================
# 🗄️ DATABASE & ORM
# ================================
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database Connection Pooling
PyMySQL==1.1.0

# ================================
# 🚀 PRODUCTION SERVER & DEPLOYMENT
# ================================
gunicorn==21.2.0
gevent==23.9.1
eventlet==0.33.3

# Process Management
supervisor==4.2.5

# ================================
# 📊 MONITORING & LOGGING
# ================================
psutil==5.9.6
prometheus-client==0.19.0
Flask-APM==0.3.0

# Advanced Logging
structlog==23.2.0
python-json-logger==2.0.7

# ================================
# 🔄 CACHING & PERFORMANCE
# ================================
redis==5.0.1
Flask-Redis==0.4.0
Flask-Caching==2.1.0

# HTTP Caching
Flask-HTTPAuth==4.8.0

# ================================
# 📧 EMAIL & NOTIFICATIONS
# ================================
Flask-Mail==0.9.1
emails==0.6
sendgrid==6.10.0

# Email Validation
email-validator==2.1.0

# ================================
# 🛡️ RATE LIMITING & SECURITY
# ================================
Flask-Limiter==3.5.0
slowapi==0.1.9

# Security Scanning
bandit==1.7.5
safety==2.3.5

# ================================
# 📈 API & DOCUMENTATION
# ================================
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
marshmallow==3.20.1
apispec==6.3.0

# API Documentation
flasgger==0.9.7.1

# ================================
# 🔧 UTILITIES & HELPERS
# ================================
python-dotenv==1.0.0
click-log==0.4.0
colorama==0.4.6

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
lxml==4.9.3
beautifulsoup4==4.12.2

# ================================
# 📸 IMAGE & FILE PROCESSING  
# ================================
Pillow==10.1.0
cairosvg==2.7.1

# SVG Processing
svglib==1.5.1

# ================================
# 🧪 TESTING & QUALITY ASSURANCE
# ================================
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Load Testing
locust==2.17.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# ================================
# 🔍 DEVELOPMENT & DEBUGGING (Optional)
# ================================
# Uncomment for development debugging in production
# flask-debugtoolbar==0.13.1
# werkzeug-profiler==0.1.1

# ================================
# 🌍 INTERNATIONALIZATION (Optional)
# ================================
# Flask-Babel==4.0.0
# Babel==2.13.1

# ================================
# 📊 ANALYTICS & METRICS (Optional)
# ================================
# google-analytics-data==0.17.1
# mixpanel==4.10.0

# ================================
# ☁️ CLOUD SERVICES (Optional)
# ================================
# boto3==1.34.0           # AWS S3 for file storage
# google-cloud-storage==2.10.0  # Google Cloud Storage

# ================================
# 🔐 ENVIRONMENT VARIABLES
# ================================
python-decouple==3.8

# ================================
# 📝 LATEX & TIKZ PROCESSING  
# ================================
# Note: LaTeX packages (texlive, lualatex, pdf2svg) must be installed at system level
# These are managed by system package manager (apt/yum), not pip

# System dependencies required:
# - texlive-full (or texlive-latex-extra + texlive-fonts-recommended)
# - texlive-luatex
# - pdf2svg
# - imagemagick (for image processing)

# ================================
# 🔧 PRODUCTION OPTIMIZATION
# ================================
# Memory profiling
memory-profiler==0.61.0

# Performance monitoring  
py-spy==0.3.14

# HTTP/2 Support
hyper==0.7.0

# Compression
flask-compress==1.14

# ================================
# 🛠️ SYSTEM ADMINISTRATION  
# ================================
# System monitoring
netifaces==0.11.0
distro==1.8.0

# Log rotation
logrotate==2.0.2

# Process monitoring
setproctitle==1.3.3
